+++ BotFramework.Tests/UnitTest1.cs
@@ -1,11 +1,9 @@
-using System.Linq;
 using System.Threading.Tasks;
 using BotFramework.Abstractions;
 using NUnit.Framework;
 using BotFramework.Clients;
 using BotFramework.HostServices;
 using BotFramework.Middleware;
-using EchoBot;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
@@ -39,24 +37,6 @@ namespace BotFramework.Tests
                   .Services.GetService<AppRunnerServiceExtensions.DebugDelegateWrapper>()!.App;
         }
 
-        [Test]
-        public void GetStaticCommands_WhenExecuted_ShouldReturnStaticCommands()
-        {
-            var commands = CommandsMiddlewareExtensions.GetStaticCommands().StaticCommandsTypes.ToList();
-
-            Assert.Contains(typeof(EchoCommand), commands);
-            Assert.Contains(typeof(HelpCommand), commands);
-        }
-
-        [Test]
-        public void ServicesShouldContainStaticCommands()
-        {
-            var cmd      = new EchoCommand(null);
-            var commands = _host.Services.GetServices(typeof(IStaticCommand)).Select(a => a.GetType()).ToList();
-
-            Assert.Contains(typeof(EchoCommand), commands);
-            Assert.Contains(typeof(HelpCommand), commands);
-        }
 
         [Test]
         public async Task Test1()
